relocate consensus core processes MassCalculator to consensus core
